backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
ollama  # Add this line
